Add an "mtls" mode for connection.
This mode support mutual TLS on the GRPC transport connection.
This "mtls" mode required a credential_file (certificate), a key_file
(encryption key for the certificate), and a trust_file (the trust
sources for the credential certificate).

These new files can be specified in a config file or on the command
line.